+++ extensions/zen-browser/src/hooks/usePinnedTabs.ts
@@ -2,5 +2,5 @@ import { usePromise } from "@raycast/utils";
 import { readBridgeTabs } from "../util/bridge";
 
 export function usePinnedTabs() {
-  return usePromise(readBridgeTabs, [], { onError: () => {} });
+  return usePromise(readBridgeTabs, [], { onError: () => {}, execute: process.platform === "darwin" });
 }